requirement.txt
asgiref==3.8.1
dj-database-url==3.0.1
Django==5.2.4
djangorestframework==3.16.0
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
PyMySQL==1.1.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
